Cleaned and refactored main
