+++ client/main.py
@@ -1,5 +1,5 @@
 import socket
-import event_pb2
+import block_device_event_pb2
 
 
 with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as client:
@@ -8,7 +8,7 @@ with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as client:
     while True:
         data = client.recv(1024)
 
-        event = event_pb2.Event()
+        event = block_device_event_pb2.BlockDeviceEvent()
         event.ParseFromString(data)
 
         print(f"Received Protobuf Event: Action={event.action}, Devname={event.devname}, Vendor={event.vendor}, Model={event.model}, Size={event.size}, Partitions={event.partitions}, Type={event.type}")
