+++ src/common/socket.hpp
@@ -18,3 +18,5 @@ public:
 private:
     const AutoFd _socket_fd;
 };
+
+std::unique_ptr<Socket> accept_client(const uint32_t port);
